SpawnerReaper: initial worker import from CppBlankProject
This is just an instance of the blank managed C++ worker from
SpatialOS's C++ template project.  I am staging this as its own commit
to demonstrate what happens when you use the worker_create.sh script.

Created by using
https://github.com/spatialos/CppBlankProject
to create a new worker into a clone of
https://github.com/keunwoo/spatialos-empty-project
as follows (from inside the CppBlankProject directory, which is a
sibling of the spatialos-empty-project directory):

bash ./worker_create.sh Managed SpawnerReaper ../spatialos-empty-project