add basic, dumb SpawnerReaper worker
This worker simply spawns "ticktock" entities in a 5x5 grid near the
origin, and then deletes them after about 10 seconds.  Hopefully this
code gives at least a little flavor of what it's like to program
against "bare metal" SpatialOS C++ APIs, without the Unity layers that
most of the public SpatialOS tutorials assume.

There are many deficiencies in the design and implementation of this
worker (read the comments for some of them), but I am pushing this up
just to share what I've done so far.